Implements interrupt based key matrix driver
The device is put to sleep with all row pins driving high and waiting for
a rising edge on the col pins. During the ISR a 40ms debounce timer is
started and the interrupt is disabled. The key is then obtained by
scanning the matrix and the device is put to sleep. This time only the
"active" row is driving and the signal is again debounced on the falling
edge. After which we must process the key press and go back to sleep.