Implements base conversion functions
A buffer stores the current numerical input as a C-String to facilitate printing to the LCD and 1:1 mapping on the keypad. Auxiliary functions were implemented to convert the C-string to an unsigned int and back again. Currently, I've limited these conversions to 16-bits to work with both the PIC's 16-bit ints and the LCD's 16 character size.